+++ packages/create-codes/templates/react/src/modules/count-context/context/useCounter.ts
